some message here

+++ database/migrations/2024_02_23_091135_create_animals_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('animals', function (Blueprint $table) {
             $table->id();
-            $table->integer('image_id')->nullable;
+            $table->bigInteger('image_id')->nullable;
             $table->foreignId('owner_id')->nullable;
             $table->string('name') ->nullable;
             $table->string('species') ->nullable;
